requirements.txt
python-dotenv
aiogram
aiosqlite
loguru
beautifulsoup4
requests